Formula vector

This function defines a new modified formula class that has been vectorized. It expands upon the functionality of formulas.

Usage

formula_rx(x = unspecified(), ...)

# S3 method for term_rx
formula_rx(
  x,
  roles = list(),
  groups = list(),
  labels = list(),
  pattern = character(),
  ...
)

# S3 method for formula
formula_rx(
  x,
  roles = list(),
  groups = list(),
  labels = list(),
  pattern = character(),
  ...
)

# S3 method for default
formula_rx(x, ...)

frx(x = unspecified(), ...)

Arguments

x

Objects of the following types can be used as inputs

term_rx

...

Arguments to be passed to or from other methods

roles

Specific roles the variable plays within the formula. These are of particular importance, as they serve as special terms that can effect how a formula is interpreted. The options for roles are as below:

exposure or X(...): a predictor variable that serves as a primary or key variable in the \(Exposure ~ Outcome\) relationship

Formulas can be condensed by applying their specific role to individual terms as a function/wrapper. For example, y ~ X(x1) + x2 + x3. This would signify that x1 has the specific role of an exposure.

groups

List of formulas that have the term (or terms) on the LHS and the group name on the RHS (quotations to indicate character value not necessary). E.g. list(c(x1, x2) ~ grp)

pattern

This is the expansion pattern used to decide how the covariates will incorporated into the formulas. The options are c("direct", "sequential", "parallel"). See the details for further explanation.

direct: the covariates will all be included in each formula

sequential: the covariates will be added sequentially, one by one, or by groups, as indicated

parallel: the covariates or groups of covariates will be placed in parallel

Value

An object of class formula_rx

Details

Patterns

The expansion pattern allows for instructions on how the covariates should be included in different formulas. Below, assuming that x1, x2, and x3 are covariates...

$$y ~ x1 + x2 + x3$$

Direct:

$$y ~ x1 + x2 + x3$$

Seqential:

$$y ~ x1$$ $$y ~ x1 + x2$$ $$y ~ x1 + x2 + x3$$

Parallel:

$$y ~ x1$$ $$y ~ x2$$ $$y ~ x3$$